Restructure methods in Unifier into sibling modules
