+++ compiler/hash-typecheck/src/ops/mod.rs
@@ -3,7 +3,9 @@
 //! Code from this module is to be used while traversing and typing the AST, in order to unify
 //! types and ensure correctness.
 pub mod building;
+pub mod params;
 pub mod reader;
-pub mod scope;
+pub mod simplify;
 pub mod substitute;
+pub mod typing;
 pub mod unify;
